Replace ask_y_n() with ask_yes_no().
The new version localizes the default value, and expects localized input from
the use, when a translation is available. Its second argument, when passed,
determines whether the default is affirmative (yes) or negative (no). The user
may input a substring of either answer, such as "y" for "Yes" or "j" for "Ja".
Translators therefore must be sure that the first char of each answer is
distinct, at least for now.

The ask_y_n() method is deprecated with a warning, and its use in Sqitch
itself replaced with calls to `ask_yes_no(). This change required changing the
way MockOutput handles mock prompts, as well.

Based on a suggestion by @tiguchi in #449.